Move error types into RError file
