+++ Stingray/APIs/APIBasicNetwork.swift
@@ -45,105 +45,6 @@ public enum NetworkRequestType: String {
     case delete = "DELETE"
 }
 
-public enum NetworkError: RError {
-    /// The request URL was invalid.
-    case invalidURL(String)
-    /// Could not encode JSON.
-    case encodeJSONFailed(Error)
-    /// Could not send the payload
-    case requestFailedToSend(Error)
-    /// Response was bad in some way
-    case badResponse(responseCode: Int, response: String?)
-    /// Could not decode the returned JSON
-    case decodeJSONFailed((any RError)?, url: URL?)
-    /// An access token is needed
-    case missingAccessToken
-    
-    public var next: (any RError)? {
-        switch self {
-        case .decodeJSONFailed(let error, _): return error
-        default: return nil
-        }
-    }
-    
-    public var errorDescription: String {
-        switch self {
-        case .invalidURL(let description):
-            return "The requested URL was invalid: \(description)"
-        case .encodeJSONFailed:
-            return "Unable to encode JSON"
-        case .requestFailedToSend:
-            return "Request failed to send"
-        case .badResponse(let code, let text):
-            return "Received a bad response from the server - \(code) \(text ?? "")"
-        case .decodeJSONFailed(_, let url):
-            return "Failed to decode JSON from \(url?.absoluteString ?? "an unknown URL")"
-        case .missingAccessToken:
-            return "An access token is needed"
-        }
-    }
-}
-
-public enum JSONError: RError {
-    /// Denotes a missing entry in a given JSON object. First `String` denotes the key, and the second `String` denotes the object's name
-    case missingKey(String, String)
-    /// Denotes a missing JSON object within another JSON object. First `String` denotes the key,
-    /// and the second `String` denotes the object's name
-    case missingContainer(String, String)
-    /// Failed to decode JSON at all. The `String` denotes the object's name, `Error` is the thrown JSON error
-    case failedJSONDecode(String, Error)
-    /// Failed to encode JSON at all. The `String` denotes the object's name
-    case failedJSONEncode(String)
-    /// The unwrapped key is an unexpected value.
-    case unexpectedKey(RError)
-    
-    public var next: (any RError)? {
-        switch self {
-        case .unexpectedKey(let err): return err
-        default: return nil
-        }
-    }
-    
-    public var errorDescription: String {
-        switch self {
-        case .missingKey(let key, let objectName):
-            return "The key \(key) was missing from the JSON object \(objectName)"
-        case .missingContainer(let containerName, let parentObjectName):
-            return "The JSON object \(containerName) was missing from the JSON object \(parentObjectName)"
-        case .failedJSONDecode(let objectName, let error):
-            if let decodingError = error as? DecodingError {
-                switch decodingError {
-                case .keyNotFound(let key, let context):
-                    let path = context.codingPath.map { $0.stringValue }.joined(separator: " -> ")
-                    return "Unable to decode JSON for \(objectName): Missing key '\(key.stringValue)' at \(path)"
-                case .valueNotFound(let type, let context):
-                    let path = context.codingPath.map { $0.stringValue }.joined(separator: " -> ")
-                    return "Unable to decode JSON for \(objectName): Missing value of type '\(type)' at \(path)"
-                case .typeMismatch(let type, let context):
-                    let path = context.codingPath.map { $0.stringValue }.joined(separator: " -> ")
-                    return """
-                        Unable to decode JSON for \(objectName): Type mismatch for '\(type)' at \(path). \
-                        \(context.debugDescription)
-                        """
-                case .dataCorrupted(let context):
-                    let path = context.codingPath.map { $0.stringValue }.joined(separator: " -> ")
-                    return """
-                        Unable to decode JSON for \(objectName): Data corrupted at \(path). \
-                        \(context.debugDescription)
-                        """
-                @unknown default:
-                    return "Unable to decode JSON  \(objectName): \(error.localizedDescription)"
-                }
-            }
-            return "JSON failed to decode for \(objectName)"
-        case .failedJSONEncode(let objectName):
-            return "Failed to encode JSON for \(objectName)"
-        case .unexpectedKey:
-            return "The unwraped JSON value was unexpected"
-        }
-    }
-}
-
 /// A Jellyfin specific basic network struct for making network requests
 public final class JellyfinBasicNetwork: BasicNetworkProtocol {
     var address: URL
